+++ frontend/src/Components/Forms/Units/BasicInputField.js
@@ -1,34 +1,26 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { FloatingLabel, Form, InputGroup, Button } from "react-bootstrap";
 import FieldError from './FieldError';
 import Icon from '../../UI/Icon';
 import Input from './Input';
 import Select from "react-select";
+import { useSignupContext } from '../../../context/SignupContext';
+import { userTypeOptions, customStyles } from '../../Services/appdata';
 
-const userTypeOptions = [
-    { value: 1, label: 'Organization' },
-    { value: 2, label: 'Teacher' },
-    { value: 3, label: 'Student' }
-];
 
-const BasicInputField = ({ state, onBlurHandler, onFocusHandler, valueChangeHandler, togglePasswordVisibility, userTypeHandler }) => {
+const BasicInputField = () => {
 
-    const { name, email, password, confirmPassword, userType } = state;
+    // const { name, email, password, confirmPassword, userType } = state;
+    const { name, email, password, confirmPassword, userType, onBlurHandler, onFocusHandler, valueChangeHandler, userTypeHandler } = useSignupContext();
+    const [seePassword, setSeePassword] = useState("password");
 
-    const customStyles = {
-        control: (baseStyles, state) => ({
-            ...baseStyles,
-            borderRadius: '1rem',
-            paddingLeft: '0.625rem',
-            paddingBottom: '0.625rem',
-            paddingTop: '0.625rem',
-            width: '100%',
-            color: '#1C1C1C',
-            appearance: 'none',
-            outline: 'none',
-            boxShadow: 'none',
-        })
-    };
+    const togglePasswordVisibility = (e) => {
+        if (seePassword === "password") {
+            setSeePassword("text");
+        } else {
+            setSeePassword("password");
+        }
+    }
 
     return (
         <>
@@ -64,7 +56,7 @@ const BasicInputField = ({ state, onBlurHandler, onFocusHandler, valueChangeHand
                 <InputGroup>
                     <FloatingLabel as={InputGroup} label="Password">
                         <Form.Control
-                            type={password.type}
+                            type={seePassword}
                             name="password"
                             value={password.value}
                             placeholder="Password"
@@ -74,8 +66,8 @@ const BasicInputField = ({ state, onBlurHandler, onFocusHandler, valueChangeHand
                             className="rounded-start-4 outline-end-0 border-end-0"
                             aria-describedby="basic-addon1"
                         />
-                        <Button className="rounded-end-4 bg-white border-start-0" style={{borderColor: "lightgrey"}} id="button-addon1" onClick={togglePasswordVisibility}>
-                            {password.type === "password" ? (
+                        <Button className="rounded-end-4 bg-white border-start-0" style={{ borderColor: "lightgrey" }} id="button-addon1" onClick={togglePasswordVisibility}>
+                            {seePassword === "password" ? (
                                 <Icon name="EyeOff" color="#1c1c1c" size="18" />
                             ) : (
                                 <Icon name="Eye" color="#1c1c1c" size="18" />
@@ -89,7 +81,7 @@ const BasicInputField = ({ state, onBlurHandler, onFocusHandler, valueChangeHand
             <div className="mb-2">
                 <Input
                     label="Confirm Password"
-                    type={password.type}
+                    type={seePassword}
                     name="confirmPassword"
                     value={confirmPassword.value}
                     placeholder="Confirm Password"
